feat: create Exceptions handling

+++ src/main/java/com/mathiassammer/dscrudclientes/services/ClientService.java
@@ -3,13 +3,16 @@ package com.mathiassammer.dscrudclientes.services;
 import com.mathiassammer.dscrudclientes.dto.ClientDTO;
 import com.mathiassammer.dscrudclientes.entities.Client;
 import com.mathiassammer.dscrudclientes.repositories.ClientRepository;
+import com.mathiassammer.dscrudclientes.services.exceptions.DatabaseException;
+import com.mathiassammer.dscrudclientes.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
 
 @Service
 public class ClientService {
@@ -25,8 +28,10 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id){
-        Optional<Client> client = clientRepository.findById(id);
-        return new ClientDTO(client.get());
+        Client client = clientRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado")
+        );
+        return new ClientDTO(client);
     }
 
     @Transactional
@@ -39,15 +44,26 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto) {
-        Client entity = clientRepository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        Client saved = clientRepository.save(entity);
-        return new ClientDTO(saved);
+        try {
+            Client entity = clientRepository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            Client saved = clientRepository.save(entity);
+            return new ClientDTO(saved);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        clientRepository.deleteById(id);
+        if (!clientRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            clientRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ClientDTO dto, Client entity) {
